backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pymongo==4.10.1
dnspython==2.7.0
bcrypt==4.2.1
pydantic==2.13.3
pydantic[email]==2.13.3
pydantic-core==2.46.3
fastapi-mail==1.6.2
aiosmtplib==5.1.0
qrcode[pil]==8.0
httpx==0.28.1
python-dotenv==1.0.1
stripe==11.4.0
pyotp==2.9.0
slowapi==0.1.9
pytest==8.3.4
pytest-asyncio==0.24.0
anyio[trio]==4.7.0
